refactor: split EarningsService into aggregation and export services
Resolves #336

Improves modularity by separating concerns:
- EarningsAggregationService: aggregation, dashboard, leaderboard, platform analytics
- EarningsExportService: CSV export functionality
- EarningsService: facade that delegates to both services
- Maintains backward compatibility with existing API
